refactor: move to mono mapping file
