+++ subgraphs/smartgarden-optimism-gauges/abis/VelodromeVoterV2.json
