+++ subgraphs/smartgarden-optimism-gauges/subgraph.yaml
@@ -3,11 +3,11 @@ schema:
   file: ./schema.graphql
 dataSources:
   - kind: ethereum
-    name: VelodromeVoter
+    name: VelodromeVoterV2
     network: optimism
     source:
       address: "0x41C914ee0c7E1A5edCD0295623e6dC557B5aBf3C"
-      abi: VelodromeVoter
+      abi: VelodromeVoterV2
       startBlock: 105896851
     mapping:
       kind: ethereum/events
@@ -15,14 +15,14 @@ dataSources:
       language: wasm/assemblyscript
       entities: []
       abis:
-        - name: VelodromeVoter
-          file: ./abis/VelodromeVoter.json
+        - name: VelodromeVoterV2
+          file: ./abis/VelodromeVoterV2.json
         - name: ERC20
           file: ./abis/ERC20.json
       eventHandlers:
         - event: GaugeCreated(indexed address,indexed address,indexed address,address,address,address,address,address)
           handler: handleGaugeCreated
-      file: ./src/velodrome-voter.ts
+      file: ./src/mapping.ts
 templates:
   - kind: ethereum/contract
     name: Gauge
@@ -33,7 +33,7 @@ templates:
       kind: ethereum/events
       apiVersion: 0.0.7
       language: wasm/assemblyscript
-      file: ./src/gauge.ts
+      file: ./src/mapping.ts
       entities: []
       abis:
         - name: Gauge
